feat: exec route partially working
